+++ src/pages/MarketResearch.js
@@ -1,23 +1,9 @@
-import React, { useState, useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-import {
-    Box,
-    Flex,
-    Text,
-    VStack,
-    Tabs,
-    TabList,
-    TabPanels,
-    Tab,
-    TabPanel
-} from '@chakra-ui/react';
-import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
-import { Pie } from 'react-chartjs-2';
+import React, { useState } from 'react';
+import { Flex, Box, Text, VStack, HStack, Badge, Tabs, TabList, TabPanels, Tab, TabPanel } from '@chakra-ui/react';
 import PageLayout from '../component/common/PageLayout';
 import DateRangePicker from '../component/common/DateRangePicker';
-import PolicyTable from '../component/common/PolicyTable';
-
-ChartJS.register(ArcElement, Tooltip, Legend);
+import NewsContent from '../component/common/NewsContent';
+import ResultTable from '../component/common/ResultTable';
 
 const getTodayString = () => {
     const today = new Date();
@@ -27,75 +13,55 @@ const getTodayString = () => {
     return `${year}-${month}-${day}`;
 };
 
-const MarketResearchPage = () => {
+const MarketResearch = () => {
     const [startDate, setStartDate] = useState(getTodayString());
     const [endDate, setEndDate] = useState(getTodayString());
-    const [searchQuery, setSearchQuery] = useState('');
-    const location = useLocation();
 
-    useEffect(() => {
-        const params = new URLSearchParams(location.search);
-        const query = params.get('query');
-        if (query) {
-            setSearchQuery(query);
-            // 여기에서 검색어를 사용하여 필요한 데이터를 불러오는 로직을 구현할 수 있습니다.
-            console.log("검색어:", query);
-        }
-    }, [location]);
-
-    const policyData = [
+    const positiveData = [
         { policy: "동물 어쩌구 법", result: "긍정", count: 506 },
         { policy: "식품 법", result: "긍정", count: 500 },
         { policy: "아동 법", result: "긍정", count: 230 },
         { policy: "그 밖의 안전4법", result: "긍정", count: 120 },
     ];
 
-    const pieChartData = {
-        labels: ['긍정', '부정', '기타'],
-        datasets: [
-            {
-                data: [83.6, 11.4, 5],
-                backgroundColor: ['#3182CE', '#E53E3E', '#ECC94B'],
-            },
-        ],
-    };
+    const negativeData = [
+        { policy: "환경 규제법", result: "부정", count: 300 },
+        { policy: "세금 인상법", result: "부정", count: 250 },
+    ];
+
+    const etcData = [
+        { policy: "기타 정책 1", result: "기타", count: 150 },
+        { policy: "기타 정책 2", result: "기타", count: 100 },
+    ];
 
     return (
         <PageLayout>
-            <VStack spacing={6} align="stretch">
-                <DateRangePicker
-                    startDate={startDate}
-                    endDate={endDate}
-                    onStartDateChange={(e) => setStartDate(e.target.value)}
-                    onEndDateChange={(e) => setEndDate(e.target.value)}
-                />
+            <DateRangePicker
+                startDate={startDate}
+                endDate={endDate}
+                onStartDateChange={(e) => setStartDate(e.target.value)}
+                onEndDateChange={(e) => setEndDate(e.target.value)}
+            />
 
-                {searchQuery && (
-                    <Box>
-                        <Text fontSize="lg" fontWeight="bold">검색어: "{searchQuery}"</Text>
-                    </Box>
-                )}
+            <Flex gap={6}>
+                {/* 왼쪽: 미디어 데이터 */}
+                <Box width="40%" bg="white" borderRadius="lg" p={6} boxShadow="md">
+                    <Text fontSize="xl" fontWeight="bold" mb={4}>[미디어 데이터] 뉴스 및 미디어 데이터 분석 결과</Text>
+                    <VStack align="stretch" spacing={4}>
+                        <HStack>
+                            <Badge colorScheme="blue" p={2} borderRadius="full">뉴스 분고</Badge>
+                            <Text fontWeight="bold">이디야스</Text>
+                            <Text fontSize="sm" color="gray.500">24년7월9일 발행, 15분 전</Text>
+                        </HStack>
+                        <NewsContent
+                            title="뉴스 제목"
+                            content="뉴스 내용 어쩌구 저쩌구..."
+                        />
+                    </VStack>
+                </Box>
 
-                <Flex gap={6}>
-                    {/* 왼쪽: 정책 데이터 */}
-                    <Box width="50%" bg="white" borderRadius="lg" p={6} boxShadow="md">
-                        <Text fontSize="xl" fontWeight="bold" mb={4}>[정책 데이터] 공공 데이터 및 K-Startup 분석 결과</Text>
-                        <Text fontWeight="bold" mb={2}>정책명: {searchQuery || "어쩌구"}</Text>
-                        <Text fontSize="sm" color="gray.500">미디어데이터 24년07월10일 갱신</Text>
-                        <Text mt={2}>정책 내용 어쩌구 머쩌구 설명...</Text>
-                    </Box>
-
-                    {/* 오른쪽: 긍부정 비율 */}
-                    <Box width="50%" bg="white" borderRadius="lg" p={6} boxShadow="md">
-                        <Text fontSize="xl" fontWeight="bold" mb={4}>[긍부정 비율]</Text>
-                        <Box height="300px">
-                            <Pie data={pieChartData} options={{ responsive: true, maintainAspectRatio: false }} />
-                        </Box>
-                    </Box>
-                </Flex>
-
-                {/* 정책 테이블 */}
-                <Box bg="white" borderRadius="lg" p={6} boxShadow="md">
+                {/* 오른쪽: 분석 결과 */}
+                <Box width="60%" bg="white" borderRadius="lg" p={6} boxShadow="md">
                     <Tabs>
                         <TabList mb={4}>
                             <Tab>긍정</Tab>
@@ -105,31 +71,48 @@ const MarketResearchPage = () => {
 
                         <TabPanels>
                             <TabPanel p={0}>
-                                <PolicyTable data={policyData} />
+                                <ResultTable data={positiveData} />
                             </TabPanel>
                             <TabPanel p={0}>
-                                {/* Add negative policy data here */}
-                                <Text>부정적인 정책 데이터가 여기에 표시됩니다.</Text>
+                                <ResultTable data={negativeData} />
                             </TabPanel>
                             <TabPanel p={0}>
-                                {/* Add other policy data here */}
-                                <Text>기타 정책 데이터가 여기에 표시됩니다.</Text>
+                                <ResultTable data={etcData} />
                             </TabPanel>
                         </TabPanels>
                     </Tabs>
                 </Box>
+            </Flex>
 
-                {/* 분석 결과 요약 */}
-                <Box bg="white" borderRadius="lg" p={6} boxShadow="md">
-                    <Text fontSize="xl" fontWeight="bold" mb={4}>[분석 결과 요약]</Text>
-                    <Text>
-                        여기에 분석 결과 요약 내용이 들어갑니다. 예를 들어, 전체적인 정책 동향,
-                        주요 긍정/부정 요인, 향후 전망 등을 포함할 수 있습니다.
-                    </Text>
-                </Box>
-            </VStack>
+            {/* 분석 결과 요약 */}
+            <Box bg="white" borderRadius="lg" p={6} mt={6} boxShadow="md">
+                <Text fontSize="xl" fontWeight="bold" mb={4}>[분석 결과 요약]</Text>
+                <VStack align="stretch" spacing={4}>
+                    <Box>
+                        <Text fontWeight="bold" mb={2}>전체 분석 결과</Text>
+                        <Text>총 1,356건의 뉴스 기사를 분석한 결과, 긍정적인 평가가 56%, 부정적인 평가가 32%, 중립적인 평가가 12%로 나타났습니다.</Text>
+                    </Box>
+                    <Box>
+                        <Text fontWeight="bold" mb={2}>주요 긍정적 요인</Text>
+                        <Text>1. 동물 보호법 개정으로 인한 동물 복지 향상</Text>
+                        <Text>2. 식품안전법 강화로 소비자 신뢰도 상승</Text>
+                        <Text>3. 아동 보호 정책 확대로 사회 안전망 강화</Text>
+                    </Box>
+                    <Box>
+                        <Text fontWeight="bold" mb={2}>주요 부정적 요인</Text>
+                        <Text>1. 환경 규제법 강화로 인한 기업 부담 증가</Text>
+                        <Text>2. 세금 인상에 대한 시민들의 불만</Text>
+                    </Box>
+                    <Box>
+                        <Text fontWeight="bold" mb={2}>향후 제언</Text>
+                        <Text>1. 긍정적 평가를 받은 정책들의 지속적인 모니터링 및 개선</Text>
+                        <Text>2. 부정적 평가를 받은 정책에 대한 보완책 마련</Text>
+                        <Text>3. 중립적 평가를 받은 정책들에 대한 추가 홍보 및 설명 필요</Text>
+                    </Box>
+                </VStack>
+            </Box>
         </PageLayout>
     );
 };
 
-export default MarketResearchPage;
+export default MarketResearch;
